# api/src/main/resources/static/openapi.yml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost:8080
  description: Generated server url
paths:
...
  /api/v1/states/{stateId}/cities:
    get:
      tags:
      - state-controller
      operationId: getCitiesByState
      parameters:
      - name: stateId
        in: path
        required: true
        schema:
          type: integer
          format: int32
      responses:
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/BaseResponseListCityResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        validations:
          type: array
          items:
            $ref: "#/components/schemas/ValidationError"
...
    ValidationError:
      type: object
      properties:
        field:
          type: string
        message:
          type: string
...
    BaseResponseListCityResponse:
      type: object
      properties:
        success:
          type: boolean
        message:
          type: string
        result:
          type: array
          items:
            $ref: "#/components/schemas/CityResponse"
        error:
          $ref: "#/components/schemas/ErrorResponse"
    CityResponse:
      type: object
      properties:
        id:
          type: integer
          format: int32
        name:
          type: string
        stateId:
          type: integer
          format: int32
...